chore(routers): split test_job_manager
